fix: Modify `read_pipeline_config_from_yaml` logic of reading the env variables
With the current modification, we will check the YAML file param variables values instead of containing all the environment variables. If the param variable is env_variable with YAML configuration style, the function of the process is to get the variable value from the environment; otherwise, it will return the same YAML value if not.